+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### For Developers
 
+ - Added `GetTarget` and `ClearTarget` methods to the API
 
 ### Known Issues